Add files via upload
For file1.py, the first code there gets three websites, 1 international and 2 local(Nigerian) and has a program that fetches the daily news headlines and links to the stories from these websites. The second code there writes a program that downloads all the images in a a given website.

For file2.py, the first code is "HTTP Request in Python (Creating Sockets)". The second code is making use of BeautifulSoup for WebScraping.

For file3.py, the first code is "Trying to use the seek() and tell() functions to locate positions of file ". The second code is: 
"Creating a tool that manages files and directories, including writing functions. Create a file search function to search for files within directories and sub directories based on name or file type. Analyze file type and provide statistics on type distribution. Navigate through directories by listing files within directories, changing and creating new directories. Implement function to carry out operations like moving, copying, renaming and deleting files".

For file4.py, the both codes are writing a python program that checks every line in a file and returns the email address.